bilbo contribution!! Mummi contribution!! veriverse contribution!! LCTV contribution!! veriverse contribution!! LCTV great contribution!! catherinacortega, davidwisser, drmjg (Michael Jurgen Garbade), frankly444, fuel (Benjamin Kuhl), intelijens, vapaamies (Vladislav Javadov), linzeming, denvercowboy and qoooeo contribution!! Twitch.tv contribution!! Casey Muratori and Handmade group!! Beam.pro contribution!! J4Wx and jordanlee833 contribution!! kranalex contribution!! bidzapfc, intelijens, pablosj (Pawel Kulesza), r00tkiller (Alexandre) contribution!! catherina is back contribution!! Waiting for fuel contribution (bit anxious in fact...) !! bugsbedamned, dopedopedope, fibofx, fork, marianyc, mushmouth, onlylinuxstuff and polandball contribution!! hbc, hudwan, joshua911, justlpd, marianyc, mushmouth (Patrick McFarland), pudottapommin, sesom and wangdian2663 contribution!! Twitch contribution!! Beam contribution!! Carol and Lucinho contribution!!
Thank you Carlos!!

Thank you Mummi!!

Carol contribution!!

at bilbo at veriverse!!
i7 6700 16GB RAM SSD OCZ 120GB SSD 240GB HD 2TB Windows 10 Pro

+++ appseed/base/user/simple/simple_scroll_bar.cpp
@@ -429,7 +429,7 @@ int32_t simple_scroll_bar::SetTrackingPos(point point)
 
       dRate = MIN(1.0, dRate);
 
-      nPos = dRate * (m_scrollinfo.nMax - m_scrollinfo.nMin - m_scrollinfo.nPage);
+      nPos = (int32_t) (dRate * (m_scrollinfo.nMax - m_scrollinfo.nMin - m_scrollinfo.nPage));
 
       nPos += m_scrollinfo.nMin;
